requirements.txt
PyMuPDF==1.24.14
Pillow==11.0.0
tkinterdnd2==0.4.2
pyinstaller==6.16.0